rest сервис

+++ mobile_service/src/main/java/com/example/mobile/repos/UserRepos.java
@@ -3,5 +3,5 @@ package com.example.mobile.repos;
 import com.example.mobile.domain.Users;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepos extends CrudRepository<Users, Long> {
+public interface UserRepos extends CrudRepository<Users, Integer> {
 }
